Fix issue with linking to a material from another blend file that includes ramps.
The problem is linking to the material doesn't include the blender ramp nodes from our fake node group. To fix this:

* Create a CollectionProperty
for our ramp nodes to hold
the knot values and positions.
* On scene save, we populate
the collection properties with
the ramp values
* Add a new button to the panels
whenever we can't find the node group for the ramps. The user
will have to push this button to fix ramps.
* When the button is pushed, we
create the necessary node group and nodes for the ramps, and
copy the values from the CollectionProperty
* On scene load, we update
the ramp nodes that are from
linked blender files, in case they have been changed in the linked file.