Some fixes for ramps:
* Add a rman_fake_group_ptr property to nodes with
ramps. This is a pointer to the node group with the ramps. We shouldn't rely on the name of the nodegroup
since that could change
* In rman_scene_sync, look for updates to ShaderNodeTrees. If a tree, is a nodegroup,
we update all of the users of the nodetree. This fixes
the problem where updating the ramp nodes wouldn't
trigger an update to the material
* add a version string property to the scene settings
* add an upgrade scene handler, so we can handle
updating people's ramp nodes.