For RMAN-18482. The blender display driver can now take a callback function when the buffer is updated. The callback function will then call tag_redraw to indicate the viewport needs redrawing.
Unfortunately, this only works in RIS. In XPU, we seem
to be caught in a race condition, where we're trying
to update the camera and draw in the viewport
in the same view_draw call. So for XPU, we do the same
thing as before, but we check with the display driver
to see if the buffer has actually updated before calling
tag_redraw.