Modify the prototype_key function.
Append either "-DATA" or "-OBJECT" if the name we're using is coming
from data or if it's the object's name. This fixes an issue seen with the USD kitchen scene.
A few of the empties had the same name as a couple of the mesh datablocks, which cause the RmanSceneSync to think the types didn't match and deleted the objects.