Default to using aggregates for volumes.
Since XPU only supports aggregates, we now
create a globalVolumeAggregate as a default. Unless tagged otherwise, all volumes go into
the global aggregate. If it's already in a different aggregate or the "Globa Volume Aggregate" checkbox is turned off, only
then will the volume not be included in the global aggregate.